ai_influencer/scripts/requirements.txt
albumentations==1.4.7
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
rembg[gpu]==2.0.57
onnxruntime-gpu==1.18.1
insightface==0.7.3
tqdm==4.66.5
scikit-image==0.24.0
matplotlib==3.9.2
pyyaml==6.0.2
requests==2.32.3
